Чуть изменена корзина, создал отдельную кнопку для добавления в нее
